Codegen abis from the rust side
