--- codegenerator/cli/templates/dynamic/codegen/src/Abis.res.hbs
@@ -1,21 +1,0 @@
-module EventSignatures = {
-{{#each codegen_contracts as |contract|}}
-let {{contract.name.uncapitalized}} = [
-  {{#each contract.event_signatures as |event_signature|}}
-    "{{event_signature}}",
-  {{/each}}
-  ]
-
-{{/each}}
-let all = [
-  {{#each codegen_contracts as |contract|}}
-   {{contract.name.uncapitalized}},
-  {{/each}}
-  ]->Belt.Array.concatMany
-}
-
-{{#each codegen_contracts as |contract|}}
-let {{contract.name.uncapitalized}}Abi: Js.Json.t = %raw(`
-{{contract.abi}}
-`)
-{{/each}}
